change cell type to uintptr_t

+++ forth.c
@@ -8,6 +8,7 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 // #include <unistd.h>
 // #include <sys/syscall.h>
 #include <ctype.h>
@@ -53,7 +54,7 @@
 #define STATE_IMMEDIATE 0
 #define STATE_COMPILE   1
 /* the most important type, the cell. MUST be exactly of the pointer length! */
-typedef long long cell; // todo: use long long or intptr or uintptr?
+typedef uintptr_t cell; // todo: use long long or intptr or uintptr?
 /* preprocessor trick to test sizeof(long)==sizeof(void*)? */
 
 /* dictionary definition header. NEVER change the order of these fields, it's crucial! */
